% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoML.R
\name{auto_ml_mydata}
\alias{auto_ml_mydata}
\title{Automatic Machine Learning}
\usage{
auto_ml_mydata(data)
}
\arguments{
\item{data}{A dataframe to be fed into Auto ML. Should contain the dependent variable as "dvcol" in the dataset}
}
\value{
A list with AutoML leaderboard, Variable importance plots, Variable imporatnace measures, parameters and test accuracy for all the models generated
}
\description{
The Automatic Machine Learning (AutoML) function automates the supervised machine learning model training process. The current version of AutoML trains and cross-validates a Random Forest, an Extremely-Randomized Forest, a random grid of Gradient Boosting Machines (GBMs), a random grid of Deep Neural Nets, and then trains a Stacked Ensemble using all of the models by using h2o package.
}
