% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fcreate_Prerequisites.R
\name{combinations}
\alias{combinations}
\title{Generate all combinations of the columns, taken 2 at a time}
\usage{
combinations(df)
}
\arguments{
\item{df}{Any data frame for which the combination of columns are to be generated.}
}
\value{
An array of all the possible 2 set combinations to all the column names
}
\description{
Takes in a data frame and generates all combinations of the columns of a dataframe taking 2 columns at a time.
}
