% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fcreate_Prerequisites.R
\name{cube}
\alias{cube}
\title{CUbe of a number}
\usage{
cube(x)
}
\arguments{
\item{x}{A numeric value which needs to be cubed}
}
\value{
A numeric value representing the cube of the given number
}
\description{
Takes in a numeric value and generates cube of that.
}
