% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fcreation.R
\name{fcreate}
\alias{fcreate}
\title{A consolidated function to create features for a given dataset.}
\usage{
fcreate(data, DV, depth, returnPossibilities)
}
\arguments{
\item{data}{A dataframe on which features are to be created}

\item{DV}{Dependent variable in the dataframe in order to eliminate that before creating features}

\item{depth}{depth value}

\item{returnPossibilities}{Flag indicating whether to return list of possibilities or not. Defaulted to FALSE.}
}
\value{
When returnPossibilities=TRUE, returns the set of all possible features that can be created for a given dataset. Returns the dataframe with features created otherwise.
}
\description{
This is a consolidated function to create features for a given dataset at any given depth. Check opps() and opps_gen_iter() functions for more details.
}
