% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FcreateFit.R
\name{fcreateFit}
\alias{fcreateFit}
\title{A fit file to create the feature set for test data}
\usage{
fcreateFit(dataframe, fcreated_poss, selected_cols)
}
\arguments{
\item{dataframe}{Test dataset}

\item{fcreated_poss}{A fcreated_poss object returned from fcreate() when returnPossibilities is set to TRUE}

\item{selected_cols}{A list object returned from fselect()}
}
\value{
A dataframe with created feature based on selected cols provided
}
\description{
Takes in the details of features created for train dataset and re-creates it for the test dataset
}
