% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fselect.R
\name{fselect}
\alias{fselect}
\title{Eliminate all columns in a dataframe that have high multi-collinearity}
\usage{
fselect(fcreated)
}
\arguments{
\item{fcreated}{A dataframe with features to be tested for multi-collinearity and eliminate features with high multi-collinearity}
}
\value{
Returns a list of two objects
\itemize{
\item A dataframe with all the multi-collinear columns removed
\item A fit file indicating the list of columns that are selected
}
}
\description{
Of all the features created using fcreate(), it is possible that some of the features can be multi-collinear with each other. So it is a good practice to eliminate such columns before passing the features to a machine learning model. This function eliminates all columns that have VIF that is greater than the threshold provided.
}
