% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fcreation.R
\name{opps}
\alias{opps}
\title{Performs all possible operations on a dataframe, taking 1 or 2 columns at a time}
\usage{
opps(dataframe, returnPossibilities)
}
\arguments{
\item{dataframe}{Any data frame for which the above mentioned operations are to be generated.}

\item{returnPossibilities}{Flag indicating whether to return list of possibilities or not. Defaulted to FALSE.}
}
\value{
A data frame with all the operations performed when returnPossibilities = FALSE. Return a list of all possibilities otherwise.
}
\description{
Takes in a data frame and performs all possible operations on the columns of a dataframe taking 2 columns at a time. The basic opeartions performed on the columns are as follows:
\itemize{
\item Addition
\item Multiplication
\item Division
\item Percentile
\item Cumulative Sum
\item Cumulative Mean
\item Sin
\item Cosine
\item Tan
\item Square
\item Cube
}
}
