% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fcreation.R
\name{opps_gen_iter}
\alias{opps_gen_iter}
\title{Features are created by iteratively performing all possible operations on a dataframe based on the depth value, taking 1 or 2 columns at a time}
\usage{
opps_gen_iter(dataframe, depth, returnPossibilities)
}
\arguments{
\item{dataframe}{A dataframe on which features are to be created}

\item{depth}{depth value}

\item{returnPossibilities}{Flag indicating whether to return list of possibilities or not. Defaulted to FALSE.}
}
\value{
When returnPossibilities=TRUE, returns the set of all possible features that can be created for a given dataset. Returns the dataframe with features created otherwise.
}
\description{
Feature creation happens at a depth level. If depth = 1: operations are performed on the raw data provided by the user, If depth = 2: opeartions are performed on the raw data along with the features created at depth 0 adn so on..
}
