% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoML.R
\name{plot_H2o_models}
\alias{plot_H2o_models}
\title{Plot Variable Importances}
\usage{
plot_H2o_models(vi, model_name, cols_match, num_of_features = NULL)
}
\arguments{
\item{vi}{Variable importance dataframe returned from H2O model}

\item{model_name}{A string to display as title of the plot}

\item{cols_match}{A dataframe with original and replaced colnames returned from auto_ml_my_data}

\item{num_of_features}{The number of features shown in the plot (default is 10 or all if less than 10)}
}
\description{
Plot variable importances
}
