% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fcreate_Prerequisites.R
\name{square}
\alias{square}
\title{Square of a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value which needs to be squared}
}
\value{
A numeric value representing the square of the given number
}
\description{
Takes in a numeric value and generates square of that.
}
