% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fselect.R
\name{vif_func}
\alias{vif_func}
\title{Calculate variance inflation factor (VIF) from the result and lm() and use that to eliminate the columns with high VIF}
\usage{
vif_func(in_frame, thresh, trace)
}
\arguments{
\item{in_frame}{A dataframe with feature to be tested for multi-collinearity}

\item{thresh}{A cut-off for VIF value. Defaulted to 10}

\item{trace}{A flag which when true, prints results after each iteration. Defaulted to TRUE}
}
\value{
A data frame with all the multi-collinear features eliminated
}
\description{
Of all the features created using fcreate(), it is possible that some of the features can be multi-collinear with each other. So it is a good practice to eliminate such columns before passing the features to a machine learning model. This function eliminates all columns that have VIF that is greater than the threshold provided.
}
